Remove cli and migrate e2e tests to multi dataset paradigm
